requirements.txt
pytesseract
Pillow
Requests
mss
opencv-python
discord
